refactor(tests): Refactor tests
